[Update] Fix uint32_t to size_t

+++ inc/prsice.hpp
@@ -412,8 +412,8 @@ private:
      */
     void gen_cov_matrix(const std::string& c_cov_file,
                         const std::vector<std::string>& cov_header_name,
-                        const std::vector<uint32_t>& cov_header_index,
-                        const std::vector<uint32_t>& factor_cov_index,
+                        const std::vector<size_t>& cov_header_index,
+                        const std::vector<size_t>& factor_cov_index,
                         const std::string& delim, Reporter& reporter);
     /*!
      * \brief Function use to process the covariate file, should be able to
@@ -431,9 +431,9 @@ private:
      */
     void process_cov_file(
         const std::string& cov_file,
-        const std::vector<uint32_t>& factor_cov_index,
+        const std::vector<size_t>& factor_cov_index,
         std::vector<size_t>& cov_start_index,
-        const std::vector<uint32_t>& cov_index,
+        const std::vector<size_t>& cov_index,
         const std::vector<std::string>& cov_name,
         std::vector<std::unordered_map<std::string, size_t>>& factor_levels,
         size_t& num_column, const std::string& delim, Reporter& reporter);
